--- src/app/layout-old.js
@@ -1,21 +1,0 @@
-import { Nunito } from 'next/font/google'
-import '@/app/global.css'
-
-const nunitoFont = Nunito({
-    subsets: ['latin'],
-    display: 'swap',
-})
-
-const RootLayout = ({ children }) => {
-    return (
-        <html lang="en" className={nunitoFont.className}>
-            <body className="antialiased">{children}</body>
-        </html>
-    )
-}
-
-export const metadata = {
-    title: 'SPOON.PH',
-}
-
-export default RootLayout
